#25 Moved the product category sorting into its own file
