+++ src/Components/Products/receivedProducts.ts
@@ -1,43 +1,9 @@
-import { productCategory } from "../../Utils/enums.js";
-import {
-  ProductsArray,
-  Products,
-  SupProductCategory,
-} from "../../Utils/interface.js";
+import { ProductsArray, Products } from "../../Utils/interface.js";
 
 import { buildFeaturedProducts } from "./buildFeaturedProducts.js";
-
-import {
-  beautyProducts,
-  fragrancesProducts,
-  furnitureProducts,
-  groceriesProducts,
-  homeDecorationProducts,
-  kitchenAccessoriesProducts,
-  laptopProducts,
-  mensShirtsProducts,
-  mensShoesProducts,
-  mensWatchesProducts,
-  mobileAccessoriesProducts,
-  motorcycleProducts,
-  skinCareProducts,
-  smartphoneProducts,
-  sportsAccessoriesProducts,
-  sunglassesProducts,
-  tabletsProducts,
-  topsProducts,
-  vehicleProducts,
-  womensBagsProducts,
-  womensDressesProducts,
-  womensJewelleryProducts,
-  womensShoesProducts,
-  womensWatchesProducts,
-} from "./productCategories.js";
+import { sortProducts } from "./sortProducts.js";
 
 export let allProductsArray: Products[] = [];
-let supCategoryArray: SupProductCategory[] = [];
-let mensClothes: SupProductCategory[] = [];
-let womensClothes: SupProductCategory[] = [];
 
 export const receivedProducts = async (products: ProductsArray) => {
   console.log(products);
@@ -59,222 +25,4 @@ export const receivedProducts = async (products: ProductsArray) => {
   console.log("All Products");
   console.log(allProductsArray);
   sortProducts(allProductsArray);
-};
-
-//Issue was the missing [] (this caused it to only take 1 data object and not the array of objects)
-export const sortProducts = async (products: Products[]) => {
-  products.map((product: Products) => {
-    let allProductCategories = product.category;
-    const sortProductsCallback = (allProducts: productCategory) => {
-      switch (allProducts) {
-        case productCategory.beauty:
-          beautyProducts.push(product);
-          break;
-
-        case productCategory.fragrances:
-          fragrancesProducts.push(product);
-          break;
-
-        case productCategory.skinCare:
-          skinCareProducts.push(product);
-          break;
-
-        case productCategory.furniture:
-          furnitureProducts.push(product);
-          break;
-
-        case productCategory.groceries:
-          groceriesProducts.push(product);
-          break;
-
-        case productCategory.homeDecoration:
-          homeDecorationProducts.push(product);
-          break;
-
-        case productCategory.kitchenAccessories:
-          kitchenAccessoriesProducts.push(product);
-          break;
-
-        case productCategory.laptops:
-          laptopProducts.push(product);
-          break;
-
-        case productCategory.smartphones:
-          smartphoneProducts.push(product);
-          break;
-
-        case productCategory.mobileAccessories:
-          mobileAccessoriesProducts.push(product);
-          break;
-
-        case productCategory.tablets:
-          tabletsProducts.push(product);
-          break;
-
-        case productCategory.sunglasses:
-          sunglassesProducts.push(product);
-          break;
-
-        case productCategory.mensShirts:
-          mensShirtsProducts.push(product);
-          break;
-
-        case productCategory.mensShoes:
-          mensShoesProducts.push(product);
-          break;
-
-        case productCategory.mensWatches:
-          mensWatchesProducts.push(product);
-          break;
-
-        case productCategory.tops:
-          topsProducts.push(product);
-          break;
-
-        case productCategory.womensDresses:
-          womensDressesProducts.push(product);
-          break;
-
-        case productCategory.womensShoes:
-          womensShoesProducts.push(product);
-          break;
-
-        case productCategory.womensWatches:
-          womensWatchesProducts.push(product);
-          break;
-
-        case productCategory.womensBags:
-          womensBagsProducts.push(product);
-          break;
-
-        case productCategory.womensJewellery:
-          womensJewelleryProducts.push(product);
-          break;
-
-        case productCategory.motorcycle:
-          motorcycleProducts.push(product);
-          break;
-
-        case productCategory.vehicle:
-          vehicleProducts.push(product);
-          break;
-
-        case productCategory.sportsAccessories:
-          sportsAccessoriesProducts.push(product);
-          break;
-
-        default:
-          console.error(`Product Category Not Found!`);
-          break;
-      }
-    };
-    sortProductsCallback(allProductCategories);
-  });
-
-  mensClothes = [
-    {
-      supCategoryName: "Mens Clothes",
-      subCategories: {
-        mensShirtsProducts,
-        mensShoesProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/mens-shirts/Man%20Plaid%20Shirt/thumbnail.png",
-    },
-  ];
-
-  womensClothes = [
-    {
-      supCategoryName: "Womens Clothes",
-      subCategories: {
-        topsProducts,
-        womensDressesProducts,
-        womensShoesProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/tops/Girl%20Summer%20Dress/thumbnail.png",
-    },
-  ];
-
-  supCategoryArray = [
-    {
-      supCategoryName: "Home Decoration",
-      subCategories: {
-        homeDecorationProducts,
-        furnitureProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/furniture/Annibale%20Colombo%20Sofa/thumbnail.png",
-    },
-    {
-      supCategoryName: "Skin Products",
-      subCategories: {
-        beautyProducts,
-        fragrancesProducts,
-        skinCareProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/skin-care/Vaseline%20Men%20Body%20and%20Face%20Lotion/thumbnail.png",
-    },
-    {
-      supCategoryName: "Kitchen",
-      subCategories: {
-        kitchenAccessoriesProducts,
-        groceriesProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/kitchen-accessories/Boxed%20Blender/thumbnail.png",
-    },
-    {
-      supCategoryName: "Eletronics",
-      subCategories: {
-        laptopProducts,
-        smartphoneProducts,
-        mobileAccessoriesProducts,
-        tabletsProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/laptops/Lenovo%20Yoga%20920/thumbnail.png",
-    },
-    {
-      supCategoryName: "Clothes",
-      subCategories: {
-        mensClothes,
-        womensClothes,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/womens-shoes/Calvin%20Klein%20Heel%20Shoes/thumbnail.png",
-    },
-    {
-      supCategoryName: "Accessories",
-      subCategories: {
-        sunglassesProducts,
-        mensWatchesProducts,
-        womensWatchesProducts,
-        womensBagsProducts,
-        womensJewelleryProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/sunglasses/Black%20Sun%20Glasses/thumbnail.png",
-    },
-    {
-      supCategoryName: "Sports Accessories",
-      subCategories: {
-        sportsAccessoriesProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/sports-accessories/Basketball/thumbnail.png",
-    },
-    {
-      supCategoryName: "Vehicles",
-      subCategories: {
-        vehicleProducts,
-        motorcycleProducts,
-      },
-      thumbnail:
-        "https://cdn.dummyjson.com/products/images/vehicle/300%20Touring/thumbnail.png",
-    },
-  ];
-  console.log("Sup Categories");
-  console.log(supCategoryArray);
 };
